[core] Extract the CSV format into heat/core/io/csv.py
`load_csv`, `save_csv`, `supports_pandas` and `load_csv_from_folder` move
out of io.py. The format registers itself with the registry, and owns its
own `.csv` extension constant.

`supports_pandas` is now built on `_config.is_installed`, so the module
no longer imports pandas at import time; `load_csv_from_folder` resolves
it lazily via `require_dependency`. Its body is reimplemented on
`utils.load_from_folder`, which it shared with `load_npy_from_path`.

`supports_pandas` is also added to `__all__`. It was the only one of the
four `supports_*` functions missing from it, so `ht.supports_pandas()`
raised AttributeError while its three siblings worked.
